bodhiteam/requirements.txt
amqp==5.0.6
asgiref==3.3.1
autopep8==1.5.5
billiard==3.6.4.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-cors-headers==3.7.0
django-crontab==0.7.1
django-silk==4.1.0
djangorestframework==3.12.2
docopt==0.6.2
gprof2dot==2021.2.21
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.10.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.1.5
Pillow==8.2.0
pkg_resources==0.0.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycodestyle==2.6.0
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
xlrd==1.2.0
zipp==3.4.1
